% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine-data.R
\name{combine.data.hourly}
\alias{combine.data.hourly}
\title{Combine hourly data}
\usage{
combine.data.hourly(comps = list(), tot = NA)
}
\arguments{
\item{comps}{list of components of hourly energy generated by
\code{\link{get.inf.data.hourly}}.}

\item{tot}{Total energy generated by \code{\link{get.uni.data.hourly}}}
}
\value{
data frame with class \code{hourly} containing the
components and the residual Other energy use
}
\description{
Combine hourly data
}
\author{
David Sterratt
}

