% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meter-data.R
\name{get.inf.meter.data.hourly}
\alias{get.inf.meter.data.hourly}
\title{Get data in hourly chunks from Informatics meter}
\usage{
get.inf.meter.data.hourly(from, to)
}
\arguments{
\item{from}{Date from which to collect data}

\item{to}{Date to which to collect data}
}
\value{
Data frame containing the columns \code{Time} of centre of
interval, \code{kWh} energy used in that interval in kWh.
}
\description{
Get data in hourly chunks from Informatics meter
}
\author{
David Sterratt
}

