% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/meter-data.R
\name{get.inf.meter.data}
\alias{get.inf.meter.data}
\title{Get raw cumulative data from Informatics meter}
\usage{
get.inf.meter.data(from, to)
}
\arguments{
\item{from}{Date from which to collect data}

\item{to}{Date to which to collect data}
}
\value{
Data frame containing the columns \code{time} and
\code{cumkwh} cumulative energy up to that time in kWh.
}
\description{
Get raw cumulative data from Informatics meter
}
\examples{
res <- get.inf.meter.data("2014-04-30", "2014-05-01")
with(res, plot(time[-1], diff(cumkwh), type="l"))
}
\author{
David Sterratt
}

