% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inf-data.R
\name{get.inf.single.ups.date}
\alias{get.inf.single.ups.date}
\title{Get one day's data from a UPS}
\usage{
get.inf.single.ups.date(date, ups = "forumA", power.factor = 1,
  cache = TRUE)
}
\arguments{
\item{date}{Day on which to get data}

\item{ups}{UPS from which to get data}

\item{power.factor}{Power factor from which to compute real power
from apparent power. If this \code{NA}, use the real power
supplied by the UPS}

\item{cache}{If \code{prefer}, use cached data if available, but
otherwise use source data. If \code{use}, only used cached data,
and don't try source data. If \code{none}, don't use.}
}
\value{
Table with columns \code{Time} and \code{kWh}
}
\description{
Get one day's data from an Informatics server
}
\author{
David Sterratt
}

