% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inf-data.R
\name{get.inf.ups.data.hourly}
\alias{get.inf.ups.data.hourly}
\title{Get data in hourly chunks from an informatics server}
\usage{
get.inf.ups.data.hourly(from, to, ups = "forumA", ...)
}
\arguments{
\item{from}{Date from which to collect data}

\item{to}{Date to which to collect data}

\item{upss}{Vector of upss from which to collect data}
}
\value{
Data frame containing the columns \code{Time} of centre of
interval, \code{kWh} energy used in that interval in kWh and
\code{server} the name of the server to which the row relates
}
\description{
Get data in hourly chunks from an Informatics energy log.
}
\author{
David Sterratt
}

