% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ups-data.R
\name{get.single.ups.db}
\alias{get.single.ups.db}
\title{Get data for a single UPS from database}
\usage{
get.single.ups.db(from, to, ups = "forumA")
}
\arguments{
\item{from}{Date from which to collect data}

\item{to}{Date to which to collect data}

\item{ups}{UPS from which to get data}
}
\description{
Get data for a single UPS from database
}
\author{
David Sterratt
}
