% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ups-data.R
\name{get.single.ups.hourly}
\alias{get.single.ups.hourly}
\title{Get data in hourly chunks from an informatics UPS log.}
\usage{
get.single.ups.hourly(from, to, ups = "forumA", cache = TRUE, ...)
}
\arguments{
\item{from}{Date from which to collect data}

\item{to}{Date to which to collect data}

\item{ups}{UPS from which to collect data}
}
\value{
Data frame containing the columns \code{Time} of centre of
interval, \code{kWh} energy used in that interval in kWh.
}
\description{
Get data in hourly chunks from an informatics UPS log.
}
\author{
David Sterratt
}

