% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conversions.R
\name{hourly}
\alias{hourly}
\title{Hourly representation of data}
\usage{
hourly(x)
}
\arguments{
\item{x}{\code{cumulative}, \code{halfhourly} \&c object}
}
\description{
Hourly representation of data
}
\author{
David Sterratt
}

