% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{monthly}
\alias{monthly}
\title{Monthly representation of data}
\usage{
monthly(x)
}
\arguments{
\item{x}{\code{cumulative}, \code{hourly} \&c object}
}
\description{
Monthly representation of data
}
\author{
David Sterratt
}
