% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot.daily}
\alias{plot.daily}
\title{Plot energy data daily}
\usage{
\method{plot}{daily}(dat, col = NULL, ylim = NULL, per.hour = FALSE)
}
\arguments{
\item{dat}{Data frame of daily data with first column containing Time}

\item{col}{Colours to plot non-time columns}

\item{ylim}{Specify ylim}
}
\description{
Plot energy data daily
}
\author{
David Sterratt
}

