% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot.hourly}
\alias{plot.hourly}
\title{Plot energy data hourly}
\usage{
\method{plot}{hourly}(dat, col = NULL, ylim = NULL, ylab = "kW")
}
\arguments{
\item{dat}{Data frame of hourly data with first column containing Time}

\item{col}{Colours to plot non-time columns}

\item{ylim}{Specify ylim}
}
\author{
David Sterratt
}

