% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot.weekly}
\alias{plot.weekly}
\title{Plot energy data weekly}
\usage{
\method{plot}{weekly}(dat, col = NULL, ylim = NULL, month.bars = TRUE,
  year.bars = FALSE, main = NULL, ylab = NULL, ...)
}
\arguments{
\item{dat}{Data frame of weekly data with first column containing Time}

\item{col}{Colours to plot non-time columns}

\item{ylim}{Specify ylim}
}
\author{
David Sterratt
}

