% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{weekly}
\alias{weekly}
\title{Weekly representation of data}
\usage{
weekly(x)
}
\arguments{
\item{x}{\code{cumulative}, \code{hourly} \&c object}
}
\description{
Weekly representation of data
}
\author{
David Sterratt
}
