% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-data.R
\name{combine.data.hourly}
\alias{combine.data.hourly}
\title{Combine hourly data}
\usage{
combine.data.hourly(comps = list(), tot = NULL)
}
\arguments{
\item{comps}{list of components of hourly energy in
\code{\link{hourly}} format.}

\item{tot}{Total energy generated in \code{\link{hourly}} format.}
}
\value{
data frame with class \code{hourly} containing the
  components and the residual Other energy use
}
\description{
Combine hourly data
}
\examples{
from <- "2016-09-01"
to <- "2016-09-08"
server <- get.ups.hourly(from, to, upss=c("serverL", "serverR"), power.factor=NA)
forum <-  get.ups.hourly(from, to, upss=c("forumB"), power.factor=0.9)
meter <- hourly(get.inf.meter.data(from, to))
combine <- combine.data.hourly(list(Server=server, Forum=forum), tot=meter)
plot(combine)
}
\author{
David Sterratt
}
