% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conversions.R
\name{daily}
\alias{daily}
\title{Daily representation of data}
\usage{
daily(x)
}
\arguments{
\item{x}{\code{cumulative}, \code{hourly} \&c object}
}
\description{
Daily representation of data
}
\author{
David Sterratt
}

