% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ups-data.R
\name{dump.single.ups.to.db}
\alias{dump.single.ups.to.db}
\title{Dump UPS data to database}
\usage{
dump.single.ups.to.db(from, to, ups = "forumA", dry.run = FALSE)
}
\arguments{
\item{from}{Date from which to collect data}

\item{to}{Date to which to collect data}

\item{ups}{UPS from which to get data}

\item{dry.run}{If \code{TRUE} do not dump data, but instead report
on what owould be done}
}
\description{
Dump UPS data to database
}
\author{
David Sterratt
}
