% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ups-data.R
\name{get.single.ups}
\alias{get.single.ups}
\title{Get data from a UPS}
\usage{
get.single.ups(from, to, ups = "forumA", cache = TRUE,
  method = "database", power.factor = 1)
}
\arguments{
\item{from}{Date from which to collect data}

\item{to}{Date to which to collect data}

\item{ups}{UPS from which to get data}

\item{cache}{Relevant when method is to collect from source UPS
files. See \code{\link{get.single.ups.file}}.}

\item{method}{If \code{db}, collect from the database; otherwise
from the source UPS files}

\item{power.factor}{Power factor from which to compute real power
from apparent power. If this \code{NA}, use the real power
supplied by the UPS}
}
\value{
Table with columns \code{Time} and \code{kWh}
}
\author{
David Sterratt
}

