% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ups-data.R
\name{get.ups.hourly}
\alias{get.ups.hourly}
\title{Get hourly data from the Informatics UPS logs}
\usage{
get.ups.hourly(from, to, upss = c("forumA", "forumB"), ...)
}
\arguments{
\item{from}{Date from which to collect data}

\item{to}{Date to which to collect data}

\item{upss}{Vector of UPSs from which to collect data}
}
\value{
Data frame with columns \code{Time} (the centre of the
time period) and \code{kWh} (energy use in kWh in the hour
centred on \code{Time})
}
\description{
Get hourly data from the Informatics UPS logs
}
\author{
David Sterratt
}

