% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{hourly}
\alias{hourly}
\title{Hourly representation of data}
\usage{
hourly(x)
}
\arguments{
\item{x}{\code{cumulative} or \code{halfhourly} object}
}
\value{
Object with hourly bins starting from "to" attribute of
  \code{x} and ending with "from" attribute of \code{x}
}
\description{
Hourly representation of data
}
\author{
David Sterratt
}
