% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{stepplot}
\alias{stepplot}
\title{Step plot}
\usage{
stepplot(x, y, xlab = "Time", ylab = "kW", col = "white", ylim = NULL,
  ...)
}
\arguments{
\item{x}{Centres of data}

\item{y}{Matrix of heights of data at centres. One row per set of
data.}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{col}{Vector of colours to plot each set of data.}
}
\description{
Step plot
}
\author{
David Sterratt
}

